Add fan-network check for db_router_address method
There are cases where the mysql-router maybe running in a machine with
both a fan-network and a common interface and both will be candidates
for the db-router relation.

If the mysql unit is running on top of a container with only fan-net
available, then the other mysql-router units need to use the fan-net
IP as well, including to set the permission accesses in the database.

This change checks which IP should be used from mysql-router to connect
with mysql database via "ip route get" and sets that IP as the default.

Closes-bug: #1940843